+++ src/parser.js
@@ -204,6 +204,7 @@ var grammar = {
 
         property: [
             ['SYMBOL', code`$1: fn.get(current, "$1")`],
+            ['$ SYMBOL', code`$2: $$2`],
             ['SYMBOL : e', code`$1: $3`],
             ['[ e ] : e', code`[$2]: $5`],
             ['...', code`...current`],